minor gfx issues
git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@234 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/dialog/CTextBox.cpp
@@ -9,7 +9,7 @@
 #include "../CGraphics.h"
 #include "../StringUtils.h"
 
-CTextBox::CTextBox(float x, float y, float w, float h, const std::string& text, bool border_rel) {
+CTextBox::CTextBox(int x, int y, int w, int h, const std::string& text, bool border_rel) {
 	m_fontwidth = 0;
 	m_fontheight = 0;
 	m_x = x;
@@ -20,10 +20,10 @@ CTextBox::CTextBox(float x, float y, float w, float h, const std::string& text, 
 	m_border_relative = border_rel;
 }
 
-void CTextBox::setFontDimensions(float width, float height)
+void CTextBox::setFontDimensions(int width, int height)
 {
 	m_fontwidth = width;
-	m_fontheight = (height>0.0f) ? height : width; // 0 it is square
+	m_fontheight = (height>0) ? height : width; // 0 it is square
 	if(m_border_relative)
 	{
 		m_x += m_fontwidth;
@@ -47,7 +47,7 @@ void CTextBox::format()
 	{
 		word = GetNextWord(it, m_String);
 
-		if((oneline.length() + word.length()) > m_w/m_fontwidth) // if the line is full
+		if((oneline.length() + word.length()) > (unsigned int) (m_w/m_fontwidth)) // if the line is full
 		{
 			if(*(oneline.end()-1) == ' ') oneline.erase(oneline.end()-1);
 			m_Textline.push_back(oneline);
@@ -74,7 +74,7 @@ void CTextBox::draw()
 	for(unsigned int i=0 ; i<m_Textline.size() ; i++ )
 	{
 		g_pGraphics->drawFont(m_Textline[i], m_x, m_y+i*m_fontheight, 0); // 0 is blank colour
-		if( i*m_fontheight > m_h )
+		if( i*m_fontheight > (unsigned int) m_h )
 			break;
 	}
 }
